Add thread-safe logging, clean up code
* Add wrapped_logger for thread-safe logging

A wrapped_logger instance is returned whenever a looger function is
called (instead of returning a regular stream).
This wrapped_logger overloads the << operator so it can act like an
output stream but instead it caches the log message internally. Once
it's done with a log message, it will lock a mutex and forward the
message to the underlying stream

* fix assignment operator in log.h

Change wrapped_logger::operator= to use a const ref as the parameter.
Change implementaion of assignment operator to checked the locked status
of the unique lock member variable of the wrapped_logger being copied
to determine whether or not to lock its own unique_lock member variable

* Clean up code

Fix issues:
* cmd_handler.cpp, removed use of 'c' variable
* utils.h, used variable 'e' in exception block
* process.cpp, replaced execvp with execve and no env.
* utils.cpp, replaced use of stat with fstream::good
* option_map.h, add assignment operator to option_map