Add resource_properties class, fix initialization
* Add resource_properties class

resource_properties is a class to represent filters.
It wraps fpga_properties C structure and is meant to be used in C++
enumeration.

* libopaec++: Initialize props_ member vars to null

Initialie resource_properties::props_ variable to nullptr in
constructors to prevent trying to free up an invalid props_ in the
destructor. NOTE: This does not set props_ in the assignment operator as
it should be left as is in case it fails to clone it using
fpgaCloneProperties.