mb1: Fix mb1 to support buffers larger than 2MB
* mb1: Change rd_line buffer to 4MB (from 2MB)

* mb1: simplify initialization of input buffer

Remove rd_lines buffer and initialize the input buffer directly.
This initialization is done one cacheline at a time so a struct called
cacheline_t is created with an array of  8 64-bit (8-byte) elements.
The input buffer casts a pointer to this type of structure to operate on
the buffer one cacheline at a time.
Each cacheline of the input buffer is initialized with the index of the
next cacheline of the buffer and starts at 1 and ends with the 0.

* Initialize input buffer with zero

Initialize the input buffer to all zeroes before setting the read ptr
chaising indices.