fpgadiag: make timeouts for nlb0 and nlb3 fatal
Previosly in nlb0 and nlb3, timeouts waiting for the fpga to signal a
"done" status resulted only in a warning being logged. The test would
continue to execute the required number of iterations. This change
makes timeouts fatal conditions that end the test sequence at the
current iteration, printing out the failed number of cachelines and
causing a non-zero process exit status.