fpgaflash: Specify flash with BDF instead of /dev/mtd
* fpgaflash: Specify flash with BDF instead of /dev/mtd

In order to be consistent with the other tools, instead
of specifying the flash to update by its /dev/mtd entry,
use the BDF to specify the PCIe card containing the
flash to update.  If no BDF is specified, and there is
only one card in the system, it will be updated.

Signed-off-by: Matthew Gerlach <matthew.gerlach@linux.intel.com>
---
v2: address feedback from Dariusz Smigiel
    - s/bfd/bdf/g
    - Raise exceptions instead of sys.exit().
    - Cleaned up a bunch of logic.
    - Refactored main() to call an exportable method.

* fpgaflash: Fix factory update

The patch fixes factory update that was broken in the
previous change.  Also changed sys.exit() to raising
of exceptions.

Signed-off-by: Matthew Gerlach <matthew.gerlach@linux.intel.com>