Tswhison/plugin to master (#840)
* Add BDX Integrated FPGA PCIe ID
* Restore a handshaking probe between MPF and fpgaPrepareBuffer().
ASE before about 2 weeks ago had no support for FPGA_BUF_PREALLOCATED.
MPF probes the capabilities of fpgaPrepareBuffer() by passing in
a request for 0 length memory and a NULL buf_addr. The expected response
is FPGA_OK with no other action.
* opae-c: plugins decide FPGA_BUF_PREALLOCATED
Document the special case in fpgaPrepareBuffer for determining
whether FPGA_BUF_PREALLOCATED is supported. When the special case
is detected in opae-c, allow it to flow down into the plugin for
an answer.