Merge multiple changes for the PAC DMA test from the Darby Creek branch (#1344)
* Patch changes from afu repo to bist (#430)

* Change DMA test to run once for bist tool (#453)

* adding new dma driver files to bist (#570)

Updating the dma driver files in the bist tool support dma pre-fetcher changes.

* Update dma source & test for klockwork (#607)

* update fpgabist dma source

* Fix missing braces. (#912)

* fpgabist: don't drop original CMAKE_CXX_FLAGS value

* Add new module to search for libtbb

* Work around what appears to be a g++ bug.

Version 5.x of g++ raises on error for goto around a variable that is used
solely inside the contained region. The error is gone in later versions
of the compiler. Gave in and moved the variable declaration.