opae.io: add to external (#1761)
opae.io provides both a Python package (python/opae.io) and a console
application (tools/extra/opae.io) that interact with a PCIe device via
libopaevfio.so. The application provides an interactive Python environment
for interacting with a device, its MMIO regions, and DMA buffers. The
package provides the underpinnings for the Python back end of the app.