fpgadiag: fix finding of entrypoint dir (#1884)
now that fpgadiag is installed as a Python package, the fpgadiag
entrypoint calls the fpgadiag module which in turns calls the underlying
executable. Because the fpgadiag module is installed in the Python
site-pacakges area, we can't rely on using dirname of __file__ anymore.
We should instead use sys.argv[0] to determine the entrypoint and then
get its directory.