fpgad: parent fix and ref count update
Fix mon_consider_device() so that it doesn't abort when there
is no parent token found. This allows monitoring the
FPGA_ACCELERATOR when the FPGA_DEVICE is not bound to a driver.

With the addition of ref counting for tokens, one should not
call fpgaDestroyToken() on a parent token gotten by
fpgaPropertiesGetParent().

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>