n6010 event log (#2185)
* tools: libboard: n6010: implement reading/parsing/printing event log

The event log on the n6010 contains state info (i.e. config/power status,
sensor values) from power-on and power-off events. The log is stored on
flash, which is exposed through sysfs by the dfl bmc driver.

Support reading, parsing, and printing these events, but add some flags
to control the verbosity of the printing.


Add a new 'fpgainfo events' subcommand to print the event log from
flash. The subcommand takes arguments to control verbosity, and which
events to print (boot (i.e. offset) and count).

Signed-off-by: Martin Hundebøll <mhu@silicom.dk>